elasticsearch: updated: added script.disable_dynamic: false
Scripting in Elasticsearch
in-requests — Also known as dynamic scripting, this is the usage of scripts within the requests/code. By default, dynamic scripting is disabled in Elasticsearch. To enable the dynamic scripting facility, it's necessary to change a script variable in the elasticsearch.yml file (which you can find in the config folder) to read as follows: script.disable_dynamic: false

By specifying false, we disable the dynamic scripting facility. After setting the script flag, you'll need to restart the cluster to put it into effect.